requirements.txt
typer
python-fsutil